Fix issue that damages my ears
I finally figured out that the smoothing applied to every rendered tone was
buggy. It used a wrong timeframe, leading to the value t for the smoothing
function being outside 0.0 - 1.0, which led to extreme values being output.

I now reworked the smoothing to not accept a timeframe at all, it will
always operate on the ends of a buffer.